Leave validation added for maternity and paternity leave

+++ src/main/java/com/paves/employee_leave_management/dto/LeaveRequestValidationDTO.java
@@ -15,6 +15,6 @@ public class LeaveRequestValidationDTO {
     private String leaveTypeId;
     private LocalDate startDate;
     private LocalDate endDate;
-    private Integer daysRequested;
+    private double daysRequested;
     private String reason;
 }
